seperating event package into : command, state package
