+++ app/handler/line.go
@@ -3,8 +3,8 @@ package handler
 import (
 	iqc "infoqerja-line/app/config"
 	crud "infoqerja-line/app/crud"
-	iqi "infoqerja-line/app/event/input"
 	iql "infoqerja-line/app/line"
+	state "infoqerja-line/app/state"
 	"infoqerja-line/app/utils"
 	constant "infoqerja-line/app/utils/constant"
 	"net/http"
@@ -93,11 +93,11 @@ func customCommandHandler(service *iql.Service, text string) {
 }
 
 // Private Method
-func customJobHandler(service *iql.Service, state, source, input string) {
+func customJobHandler(service *iql.Service, currState, source, input string) {
 	finder := &iql.JobState{
-		State: state,
+		State: currState,
 	}
-	data := iqi.BaseData{
+	data := state.BaseData{
 		SourceID: source,
 		Input:    input,
 	}
