+++ app/state/base.go
@@ -1,4 +1,4 @@
-package input
+package state
 
 // BaseData : the basec structure for data (DTO) in the event layer
 type BaseData struct {
